Move yaml related utils to a separate file
